requirements.txt
torch
torchvision
onnxruntime
numpy
pillow
